extracted out the image tag conatiner in a new file
