+++ src/components/app/details/cicdHistory/Artifacts.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef, useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import { showError, EmptyState, GenericEmptyState } from '@devtron-labs/devtron-fe-common-lib'
 import { copyToClipboard, importComponentFromFELibrary } from '../../../common'
 import { useParams } from 'react-router'
@@ -9,17 +9,15 @@ import { ReactComponent as OpenInNew } from '../../../../assets/icons/ic-open-in
 import { ReactComponent as Question } from '../../../../assets/icons/ic-help.svg'
 import docker from '../../../../assets/icons/misc/docker.svg'
 import folder from '../../../../assets/icons/ic-folder.svg'
-import Creatable from 'react-select/creatable';
 import noartifact from '../../../../assets/img/no-artifact@2x.png'
-import { ReactComponent as QuestionIcon } from '../../../v2/assets/icons/ic-question.svg'
 import Tippy from '@tippyjs/react'
 import '../cIDetails/ciDetails.scss'
 import { ArtifactType, CIListItemType, CopyTippyWithTextType, HistoryComponentType } from './types'
-import { ReactComponent as EditIcon } from '../../../../assets/icons/ic-pencil.svg'
 import { DOCUMENTATION, TERMINAL_STATUS_MAP } from '../../../../config'
 import { extractImage } from '../../service'
 import { EMPTY_STATE_STATUS } from '../../../../config/constantMessaging'
-import { CustomInput } from '../../../globalConfigurations/GlobalConfiguration'
+import {ImageTagsContainer} from './ImageTags'
+
 
 const ApprovedArtifact = importComponentFromFELibrary && importComponentFromFELibrary('ApprovedArtifact')
 
@@ -177,99 +175,7 @@ const CIProgressView = (): JSX.Element => {
             </EmptyState.Subtitle>
         </EmptyState>
     )
-}
-
-const ImageTagsContainer = ({ description, tagTexts }) => {
-    const [newDescription, setNewDescription] = useState(description)
-    const [isEditing, setIsEditing] = useState(false)
-    const [selectedTags, setSelectedTags] = useState(
-        tagTexts.map((tag) => ({ text: tag.text, isSoftDeleted: tag.isSoftDeleted })),
-    )
-
-    const handleEditClick = () => {
-        setIsEditing(!isEditing)
-    }
-
-    const handleDescriptionChange = (e) => {
-        setNewDescription(e.target.value)
-    }
-
-    const handleCancel = () => {
-        setNewDescription(description)
-        setSelectedTags(tagTexts.map((tag) => ({ text: tag.text, isSoftDeleted: tag.isSoftDeleted })))
-        handleEditClick()
-    }
-
-    const handleTagCreate = (newValue) => {
-        setSelectedTags([...selectedTags, { text: newValue, isSoftDeleted: false }])
-    }
-
-    const creatableRef = useRef(null)
-
-    return (
-        <div>
-            {!isEditing ? (
-                <div className="top mt-8 br-4 image-tags-container" style={{ display: 'flex' }}>
-                    <div className="ml-10" style={{ width: '734px' }}>
-                        <div className="mb-8 mt-8">{description}</div>
-                        <div className="dc__flex-wrap flex left">
-                            {tagTexts.map((tag, index) => (
-                                <ImageTagButton key={index} text={tag.text} isSoftDeleted={tag.isSoftDeleted} />
-                            ))}
-                        </div>
-                    </div>
-                    <EditIcon
-                        className="icon-dim-16 mt-8 ml-10 image-tags-container-edit__icon"
-                        onClick={handleEditClick}
-                    />
-                </div>
-            ) : (
-                <div className="mt-12 dc__border-top-n1 ">
-                    <div className="cn-7 mt-12 flex left">Release tags (eg. v1.0)
-                    <QuestionIcon className="icon-dim-16 fcn-6 ml-4 cursor" />
-                    </div>
-                    <div className="mb-8 mt-6">
-                        <Creatable
-                            placeholder="Type a tag and press enter"
-                            onCreateOption={handleTagCreate}
-                            ref={creatableRef}  
-                        />
-                    </div>
-                    <div className="dc__flex-wrap flex left">
-                        {selectedTags.map((tag, index) => (
-                            <ImageTagButton key={index} text={tag.text} isSoftDeleted={tag.isSoftDeleted} />
-                        ))}
-                    </div>
-                    <div className="cn-7">Comment</div>
-                    <div className="flex left flex-wrap dc__gap-8 w-100 mt-6 mb-12 ">
-                        <textarea
-                            value={newDescription} 
-                            onChange={handleDescriptionChange}
-                            className="flex left flex-wrap dc__gap-8 dc__description-textarea h-90"
-                        />
-                    </div>
-                    <div className="w-100 flex right">
-                        <button className="cta cancel h-32 lh-32" type="button" onClick={handleCancel}>
-                            Cancel
-                        </button>
-                        <button className="cta h-32 lh-32 ml-12" type="button">
-                            Save
-                        </button>
-                    </div>
-                </div>
-            )}
-        </div>
-    )
-}
-
-const ImageTagButton = ({ text, isSoftDeleted }) => {
-    const containerClassName = isSoftDeleted ? 'image-tag-button-soft-deleted mb-8 mr-8' : 'image-tag-button mb-8 mr-8'
-    return (
-        <div className={containerClassName}>
-            <span className="mr-8 ml-8 mt-2 mb-2">{text}</span>
-        </div>
-    )
-}
+} 
 
 export const CIListItem = ({ type, userApprovalMetadata, triggeredBy, children }: CIListItemType) => {
     if (type === 'approved-artifact') {
@@ -295,7 +201,7 @@ export const CIListItem = ({ type, userApprovalMetadata, triggeredBy, children }
             </div>
             {type === 'artifact' && (
                 <ImageTagsContainer
-                    description="The photo shows a team of firefighters standing in front of a building they have just successfully saved."
+                    description="Hello"
                     tagTexts={[
                         { text: 'v4.9.0', isSoftDeleted: false },
                         { text: 'django', isSoftDeleted: true },
